1. Added & tested bilinear interpolation 2. Added & tested elevation class interpolation 3. Fixed bugs in centroid calculation 4. Fixed type signatures preventing proper overrides of subroutines in Grid.hpp 5. Removed ll2xy() and xy2ll() from Proj2 (OBSOLETE) 6. Converted to use of giss::parse_enum<EnumType>(char *)

+++ slib/glint2/Grid_XY.hpp
@@ -31,10 +31,10 @@ public:
 		parameterization = Grid::Parameterization::L0;
 	}
 
-	virtual int ij_to_index(int i, int j) const
+	virtual long ij_to_index(int i, int j) const
 		{ return j * nx() + i; }
 
-	virtual void index_to_ij(int index, int &i, int &j) const
+	virtual void index_to_ij(long index, int &i, int &j) const
 	{
 		int n = nx();
 		j = index / n;
